pyproject.toml
[project]
name = "repl-nix-spaceforcedatafeed"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "feedparser>=6.0.11",
    "openai>=1.58.1",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "plyer>=2.1.0",
    "psutil>=6.1.1",
    "psycopg2-binary>=2.9.10",
    "python-dateutil>=2.9.0.post0",
    "pytz>=2024.2",
    "requests>=2.32.3",
    "scipy",
    "streamlit>=1.40.1",
    "tenacity>=9.0.0",
    "trafilatura>=1.12.2",
    "wordcloud",
]
